Time: 3 ms (72.07%), Space: 44.5 MB (94.72%) - LeetHub

+++ 0414-third-maximum-number/0414-third-maximum-number.java
@@ -1,12 +1,13 @@
 class Solution {
     public int thirdMax(int[] nums) {
-        Long first = null;
-        Long second = null;
-        Long third = null;
+
+        Integer first = null;
+        Integer second = null;
+        Integer third = null;
 
         for (int num : nums) {
 
-            // Ignore duplicates
+            // Skip duplicate numbers
             if ((first != null && num == first) ||
                 (second != null && num == second) ||
                 (third != null && num == third)) {
@@ -16,15 +17,19 @@ class Solution {
             if (first == null || num > first) {
                 third = second;
                 second = first;
-                first = (long) num;
-            } else if (second == null || num > second) {
+                first = num;
+            }
+
+            else if (second == null || num > second) {
                 third = second;
-                second = (long) num;
-            } else if (third == null || num > third) {
-                third = (long) num;
+                second = num;
+            }
+
+            else if (third == null || num > third) {
+                third = num;
             }
         }
 
-        return third == null ? first.intValue() : third.intValue();
+        return (third == null) ? first : third;
     }
 }
